Slim down MS2Service and move it to API. Slim down MS2Urls. Move FASTA loaders, custom annotation classes. Extract migrateRuns() from ProteinManager.
