--- ms2/api-src/org/labkey/api/ms2/SearchClient.java
@@ -1,29 +1,0 @@
-/*
- * Copyright (c) 2015-2019 LabKey Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.labkey.api.ms2;
-
-/**
- * User: jeckels
- * Date: Jan 9, 2007
- */
-public interface SearchClient {
-    public boolean setProxyURL(String proxyURL);
-    public void findWorkableSettings(boolean useAuthentication);
-
-    public int getErrorCode();
-    public String getErrorString();
-}
