+++ ms2/src/org/labkey/ms2/MS2ServiceImpl.java
@@ -16,46 +16,20 @@
 
 package org.labkey.ms2;
 
-import org.labkey.api.data.ContainerFilter;
-import org.labkey.api.ms2.MS2Service;
-import org.labkey.api.ms2.SearchClient;
-import org.labkey.api.data.TableInfo;
 import org.labkey.api.data.Container;
-import org.labkey.api.data.SimpleFilter;
+import org.labkey.api.data.DbScope;
+import org.labkey.api.data.SQLFragment;
+import org.labkey.api.data.SqlExecutor;
+import org.labkey.api.data.SqlSelector;
+import org.labkey.api.ms2.MS2Service;
+import org.labkey.api.protein.ProteinSchema;
 import org.labkey.api.security.User;
-import org.labkey.api.query.FieldKey;
-import org.labkey.ms2.pipeline.mascot.MascotClientImpl;
-import org.labkey.ms2.query.PeptidesTableInfo;
 import org.labkey.ms2.query.MS2Schema;
-import org.apache.logging.log4j.Logger;
-import org.apache.logging.log4j.LogManager;
 
-/**
- * User: jeckels
- * Date: Jan 9, 2007
- */
+import java.sql.SQLException;
+
 public class MS2ServiceImpl implements MS2Service
 {
-    @Override
-    public SearchClient createSearchClient(String server, String url, Logger instanceLogger, String userAccount, String userPassword)
-    {
-        if(server.equalsIgnoreCase(MS2RunType.Mascot.name()))
-            return new MascotClientImpl(url, instanceLogger, userAccount, userPassword);
-        return null;
-    }
-
-    @Override
-    public TableInfo createPeptidesTableInfo(User user, Container container)
-    {
-        return createPeptidesTableInfo(user, container, true, ContainerFilter.current(container), null, null);
-    }
-
-    @Override
-    public TableInfo createSequencesTableInfo(User user, Container container)
-    {
-        return new MS2Schema(user, container).getTable(MS2Schema.TableType.Sequences.toString());
-    }
-
     @Override
     public MS2Schema createSchema(User user, Container container)
     {
@@ -63,15 +37,69 @@ public class MS2ServiceImpl implements MS2Service
     }
 
     @Override
-    public TableInfo createPeptidesTableInfo(User user, Container container, boolean includeFeatureFk, ContainerFilter containerFilter, SimpleFilter filter, Iterable<FieldKey> defaultColumns)
+    public void migrateRuns(int oldFastaId, int newFastaId) throws SQLException
     {
-        // Go through the schema so we get metadata applied correctly
-        PeptidesTableInfo table = (PeptidesTableInfo)createSchema(user, container).getTable(MS2Schema.TableType.Peptides.name(), containerFilter, true, true);
-        if (null != filter)
-            table.addCondition(filter);
-        if (null != defaultColumns)
-            table.setDefaultVisibleColumns(defaultColumns);
-        table.setLocked(true);
-        return table;
+        SQLFragment mappingSQL = new SQLFragment("SELECT fs1.seqid AS OldSeqId, fs2.seqid AS NewSeqId\n");
+        mappingSQL.append("FROM \n");
+        mappingSQL.append("\t(SELECT ff.SeqId, s.Hash, ff.LookupString FROM " + ProteinSchema.getTableInfoFastaSequences() + " ff, " + ProteinSchema.getTableInfoSequences() + " s WHERE ff.SeqId = s.SeqId AND ff.FastaId = " + oldFastaId + ") fs1 \n");
+        mappingSQL.append("\tLEFT OUTER JOIN \n");
+        mappingSQL.append("\t(SELECT ff.SeqId, s.Hash, ff.LookupString FROM " + ProteinSchema.getTableInfoFastaSequences() + " ff, " + ProteinSchema.getTableInfoSequences() + " s WHERE ff.SeqId = s.SeqId AND ff.FastaId = " + newFastaId + ") fs2 \n");
+        mappingSQL.append("\tON (fs1.Hash = fs2.Hash AND fs1.LookupString = fs2.LookupString)");
+
+        SQLFragment missingCountSQL = new SQLFragment("SELECT COUNT(*) FROM (");
+        missingCountSQL.append(mappingSQL);
+        missingCountSQL.append(") Mapping WHERE OldSeqId IN (\n");
+        missingCountSQL.append("(SELECT p.SeqId FROM " + MS2Manager.getTableInfoPeptides() + " p, " + MS2Manager.getTableInfoRuns() + " r WHERE p.run = r.Run AND r.FastaId = " + oldFastaId + ")\n");
+        missingCountSQL.append("UNION\n");
+        missingCountSQL.append("(SELECT pgm.SeqId FROM ").append(MS2Manager.getTableInfoProteinGroupMemberships()).append(" pgm, ").append(MS2Manager.getTableInfoProteinGroups()).append(" pg, ").append(MS2Manager.getTableInfoProteinProphetFiles()).append(" ppf, ").append(MS2Manager.getTableInfoRuns()).append(" r WHERE pgm.ProteinGroupId = pg.RowId AND pg.ProteinProphetFileId = ppf.RowId AND ppf.Run = r.Run AND r.FastaId = ").appendValue(oldFastaId).append("))\n");
+        missingCountSQL.append("AND NewSeqId IS NULL");
+
+        int missingCount = new SqlSelector(ProteinSchema.getSchema(), missingCountSQL).getObject(Integer.class);
+        if (missingCount > 0)
+        {
+            throw new SQLException("There are " + missingCount + " protein sequences in the original FASTA file that are not in the new file");
+        }
+
+        SqlExecutor executor = new SqlExecutor(MS2Manager.getSchema());
+
+        try (DbScope.Transaction transaction = MS2Manager.getSchema().getScope().ensureTransaction())
+        {
+            SQLFragment updatePeptidesSQL = new SQLFragment();
+            updatePeptidesSQL.append("UPDATE " + MS2Manager.getTableInfoPeptidesData() + " SET SeqId = map.NewSeqId");
+            updatePeptidesSQL.append("\tFROM " + MS2Manager.getTableInfoFractions() + " f \n");
+            updatePeptidesSQL.append("\t, " + MS2Manager.getTableInfoRuns() + " r\n");
+            updatePeptidesSQL.append("\t, " + MS2Manager.getTableInfoFastaRunMapping() + " frm\n");
+            updatePeptidesSQL.append("\t, (");
+            updatePeptidesSQL.append(mappingSQL);
+            updatePeptidesSQL.append(") map \n");
+            updatePeptidesSQL.append("WHERE f.Fraction = " + MS2Manager.getTableInfoPeptidesData() + ".Fraction\n");
+            updatePeptidesSQL.append("\tAND r.Run = f.Run\n");
+            updatePeptidesSQL.append("\tAND frm.Run = r.Run\n");
+            updatePeptidesSQL.append("\tAND " + MS2Manager.getTableInfoPeptidesData() + ".SeqId = map.OldSeqId \n");
+            updatePeptidesSQL.append("\tAND frm.FastaId = " + oldFastaId);
+
+            executor.execute(updatePeptidesSQL);
+
+            SQLFragment updateProteinsSQL = new SQLFragment();
+            updateProteinsSQL.append("UPDATE " + MS2Manager.getTableInfoProteinGroupMemberships() + " SET SeqId= map.NewSeqId\n");
+            updateProteinsSQL.append("FROM " + MS2Manager.getTableInfoProteinGroups() + " pg\n");
+            updateProteinsSQL.append("\t, " + MS2Manager.getTableInfoProteinProphetFiles() + " ppf\n");
+            updateProteinsSQL.append("\t, " + MS2Manager.getTableInfoRuns() + " r\n");
+            updateProteinsSQL.append("\t, " + MS2Manager.getTableInfoFastaRunMapping() + " frm\n");
+            updateProteinsSQL.append("\t, (");
+            updateProteinsSQL.append(mappingSQL);
+            updateProteinsSQL.append(") map \n");
+            updateProteinsSQL.append("WHERE " + MS2Manager.getTableInfoProteinGroupMemberships() + ".ProteinGroupId = pg.RowId\n");
+            updateProteinsSQL.append("\tAND pg.ProteinProphetFileId = ppf.RowId\n");
+            updateProteinsSQL.append("\tAND r.Run = ppf.Run\n");
+            updateProteinsSQL.append("\tAND frm.Run = r.Run\n");
+            updateProteinsSQL.append("\tAND " + MS2Manager.getTableInfoProteinGroupMemberships() + ".SeqId = map.OldSeqId\n");
+            updateProteinsSQL.append("\tAND frm.FastaId = " + oldFastaId);
+
+            executor.execute(updateProteinsSQL);
+
+            executor.execute("UPDATE " + MS2Manager.getTableInfoFastaRunMapping() + " SET FastaID = ? WHERE FastaID = ?", newFastaId, oldFastaId);
+            transaction.commit();
+        }
     }
 }
