+++ ms2/src/org/labkey/ms2/protein/DefaultAnnotationLoader.java
@@ -21,6 +21,7 @@ import org.labkey.api.pipeline.CancelledException;
 import org.labkey.api.pipeline.PipeRoot;
 import org.labkey.api.pipeline.PipelineJob;
 import org.labkey.api.pipeline.PipelineService;
+import org.labkey.api.protein.ProteinAnnotationPipelineProvider;
 import org.labkey.api.util.DateUtil;
 import org.labkey.api.util.FileUtil;
 import org.labkey.api.util.NetworkDrive;
