--- ms2/src/org/labkey/ms2/protein/ProteinAnnotationPipelineProvider.java
@@ -1,37 +1,0 @@
-/*
- * Copyright (c) 2012 LabKey Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.labkey.ms2.protein;
-
-import org.labkey.api.module.Module;
-import org.labkey.api.pipeline.PipeRoot;
-import org.labkey.api.pipeline.PipelineDirectory;
-import org.labkey.api.pipeline.PipelineProvider;
-import org.labkey.api.view.ViewContext;
-
-public class ProteinAnnotationPipelineProvider extends PipelineProvider
-{
-    public static final String NAME = "ProteinAnnotation";
-
-    public ProteinAnnotationPipelineProvider(Module module)
-    {
-        super(NAME, module);
-    }
-
-    @Override
-    public void updateFileProperties(ViewContext context, PipeRoot pr, PipelineDirectory directory, boolean includeAll)
-    {
-    }
-}
